+++ src/jiratui/widgets/create_work_item/screen.py
@@ -11,7 +11,6 @@ from textual.widgets import Button, Input, Rule, Select, Static
 from jiratui.api_controller.controller import APIControllerResponse
 from jiratui.models import IssueType, Project
 from jiratui.utils.create_work_item import create_widgets_for_work_item_creation
-from jiratui.widgets.base import CustomTitle
 from jiratui.widgets.create_work_item.fields import (
     CreateWorkItemAssigneeSelectionInput,
     CreateWorkItemDescription,
@@ -75,8 +74,9 @@ class AddWorkItemScreen(Screen):
         return self.query_one('#additional_fields', expect_type=VerticalScroll)
 
     def compose(self) -> ComposeResult:
-        with Vertical():
-            yield CustomTitle(self.TITLE)
+        vertical = Vertical()
+        vertical.border_title = self.TITLE
+        with vertical:
             yield Static(
                 Text('Important: Fields marked with (*) are required.', style='italic orange')
             )